% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_vignettes.R
\name{change_in_vignette}
\alias{change_in_vignette}
\title{Change the eval argument in a vignette}
\usage{
change_in_vignette(path, TtoF = TRUE)
}
\arguments{
\item{path}{path to vignette}

\item{TtoF}{Logical. Change from \code{TRUE} to \code{FALSE} (the default)
or \code{FALSE} to \code{TRUE}.}
}
\description{
Change the eval argument in a vignette
}
