% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_vignettes.R
\name{change_in_vignettes}
\alias{change_in_vignettes}
\title{Change the eval argument in all vignettes}
\usage{
change_in_vignettes(TtoF = TRUE)
}
\arguments{
\item{TtoF}{Logical. Change from \code{TRUE} to \code{FALSE} (the default)
or \code{FALSE} to \code{TRUE}.}
}
\description{
Change the eval argument in all vignettes
}
