% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lines.R
\name{find_codeportion}
\alias{find_codeportion}
\alias{find_optschunkset}
\alias{find_openingchunks}
\alias{find_closingchunks}
\title{Find lines in scripts}
\usage{
find_codeportion(script)

find_optschunkset(script)

find_openingchunks(script)

find_closingchunks(script)
}
\arguments{
\item{script}{script}
}
\value{
The lines of interest numbers.
}
\description{
Find lines in scripts
}
