% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\name{find_vignettes}
\alias{find_vignettes}
\title{Find vignettes}
\usage{
find_vignettes(check_root = TRUE)
}
\arguments{
\item{check_root}{logical}
}
\value{
The vector of paths to vignettes in the package.
}
\description{
Find vignettes
}
